Refactoring spec to consolidate behavior
It is good practice to clean up data from tests; After all each test
should be repeatable and runnable in isolation. Most databases support
transactions to ensure that data is clean between each run. And there
are other paradigms (i.e. database_cleaner) for having a clean database.

Unfortunately, Fedora doesn't easily support this. So instead we need to
clean up the file ourselves.

+++ spec/models/thing_spec.rb
@@ -7,13 +7,15 @@ describe Thing do
   include ActiveFedora::TestSupport
   subject { Thing.new }
 
+  after(:each) do
+    subject.destroy if subject.persisted?
+  end
+
   it 'should persist to Fedora' do
     subject.save!
     expect {
       subject.reload
     }.to_not raise_error(ActiveFedora::ObjectNotFoundError)
-
-    subject.destroy
   end
 
   let(:title) { "A Specific Title" }
@@ -22,7 +24,6 @@ describe Thing do
     subject.save!
     subject.reload
     expect(subject.title).to eq(title)
-    subject.destroy
   end
 
 end
